can/python: Improved can_graph_* examples.
The gui examples that run the graph in a separate thread
have been clearly named can_graph_gui_threaded_*

Both the simple and responsive examples use common functionality and
similar approaches, where applicable.

Updated the can_graph.py so that it uses graph/utils.py for arguments
parsing, plugin loading. Graph examples can now be selected via command
line.

Graph related scripts were moved to python.
Common functionality used by graph examples was moved to python/graph.
